auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-multipart==0.0.6
jinja2==3.1.4
email-validator==2.1.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
slowapi==0.1.9
limits==3.11.0
